Adds additional verification, better serialization.
Passes all information about all players in a game on getplayers request.
Verifies that the HTTP method matches the one required for the operation.
Returns the correct method in case of mismatch.
Sets default value for uninitalized record members to the atom 'null', to
simplify processing by JavaScript clients.
Adds getplayer function, to get player information for a single player in a
game.